+++ src/main/java/com/cells/cells/hyperdensity/compacting/ItemHyperDensityCompactingCellBase.java
@@ -32,7 +32,7 @@ import appeng.api.storage.IMEInventoryHandler;
 import appeng.api.storage.channels.IItemStorageChannel;
 import appeng.api.storage.data.IAEItemStack;
 import appeng.api.storage.data.IItemList;
-import appeng.core.CreativeTab;
+import com.cells.core.CellsCreativeTab;
 import appeng.core.localization.GuiText;
 import appeng.items.contents.CellConfig;
 import appeng.items.contents.CellUpgrades;
@@ -73,7 +73,7 @@ public abstract class ItemHyperDensityCompactingCellBase extends Item implements
         setMaxStackSize(64);
         setHasSubtypes(true);
         setMaxDamage(0);
-        setCreativeTab(CreativeTab.instance);
+        setCreativeTab(CellsCreativeTab.instance);
     }
 
     @Override