+++ src/main/java/com/cells/cells/normal/compacting/ItemCompactingCellBase.java
@@ -10,7 +10,7 @@ import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.entity.player.EntityPlayer;
 
-import appeng.core.CreativeTab;
+import com.cells.core.CellsCreativeTab;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.ActionResult;
@@ -63,7 +63,7 @@ public abstract class ItemCompactingCellBase extends Item implements IInternalCo
         setMaxStackSize(64);
         setHasSubtypes(true);
         setMaxDamage(0);
-        setCreativeTab(CreativeTab.instance);
+        setCreativeTab(CellsCreativeTab.instance);
     }
 
     @Override